+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
+## [1.6.0] - 2025-12-26
+### Added
+- Added scanoss.json scanning config support.
+- Added new server-side configuration parameters for scanning tune-up.
+
 ## [1.5.2] - 2025-11-07
 ### Added
 - Added Custom Contents URL support (`SCANOSS_FILE_CONTENTS_URL`)